requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.1.2
Automat==22.10.0
backports.zoneinfo==0.2.1
certifi==2024.2.2
cffi==1.15.1
channels==4.0.0
charset-normalizer==3.3.2
constantly==15.1.0
cryptography==42.0.7
daphne==4.0.0
Django==3.2.25
django-cors-headers==4.1.0
django-environ==0.11.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.0
dnspython==2.3.0
drf-nested-routers==0.93.4
gunicorn==22.0.0
hyperlink==21.0.0
idna==3.7
importlib-metadata==6.7.0
incremental==22.10.0
minio==7.2.7
mongoengine==0.28.2
packaging==24.0
pika==1.3.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.20.0
PyJWT==2.8.0
pymongo==4.7.3
PyMySQL==1.1.0
pyOpenSSL==24.1.0
pytz==2024.1
redis==5.0.6
requests==2.31.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.4
Twisted==23.8.0
txaio==23.1.1
typing_extensions==4.7.1
urllib3==2.0.7
zipp==3.15.0
zope.interface==6.4.post0
